# Speed up tests of tasks