@valeriupredoi 2021-09-03
awesome, glad to see the GA test monitor is useful after all :grin: Bouwe man, have you run the GA monitor give us a few numbers for the speedup? Also what does the fork do now (sorry, my ignorance, not a comment about the functionality). Gonna mark the PR as needing changes but the only bit I'm interested is the timing comparison only :+1: