@bouweandela 2021-09-06
Hi V, this just removes a bunch of unnecessary tests, so for sure it's faster. But here are the numbers anyway:
This branch:
```
(esmvaltool) ✔ ~/src/esmvalgroup/esmvalcore [speedup-tests|⚑ 23] 
14:45 $ pytest tests/integration/test_task.py
================================================================================================================================== test session starts ===================================================================================================================================
platform linux -- Python 3.9.1, pytest-6.2.2, py-1.10.0, pluggy-0.13.1
rootdir: /home/bandela/src/esmvalgroup/esmvalcore, configfile: setup.cfg
plugins: monitor-1.6.2, env-0.6.2, forked-1.3.0, xdist-2.2.1, flake8-1.0.7, cov-2.11.1, metadata-1.11.0, mock-3.5.1, html-3.1.1, mypy-0.8.0, anyio-3.0.1
gw0 [20] / gw1 [20]
....................                                                                                                                                                                                                                                                               [100%]
--------------------------------------------------------------------------------------------- generated html file: file:///home/bandela/src/esmvalgroup/esmvalcore/test-reports/report.html ----------------------------------------------------------------------------------------------

----------- coverage: platform linux, python 3.9.1-final-0 -----------
Coverage HTML written to dir test-reports/coverage_html
Coverage XML written to file test-reports/coverage.xml

========================================================================================================================================== mypy ==========================================================================================================================================
Success: no issues found in 1 source file
================================================================================================================================== 20 passed in 15.94s ===================================================================================================================================
(esmvaltool) ✔ ~/src/esmvalgroup/esmvalcore [speedup-tests|⚑ 23] 
14:46 $ python tests/parse_pymon.py
Looking for tests that exceed 1GB resident memory.
List of 100 slowest tests (duration, path, name
3.82s tests/integration/test_task.py::test_run_tasks[spawn-2]
1.91s tests/integration/test_task.py::test_run_tasks[fork-2]
1.44s tests/integration/test_task.py::test_runner_uses_priority[runner1]
1.34s tests/integration/test_task.py::test_diagnostic_run_task[executable1-var0 <- "zg"]
1.33s tests/integration/test_task.py::test_diagnostic_run_task[executable2-import os\n\nprint('cow')]
1.29s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable1-diag_text1]
1.27s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable0-diag_text0]
1.20s tests/integration/test_task.py::test_diagnostic_run_task[executable0-tas = True\ntas@cow = 22\n]
1.04s tests/integration/test_task.py::test_run_tasks[fork-15]
0.93s tests/integration/test_task.py::test_run_tasks[fork-None]
0.38s tests/integration/test_task.py::test_py2ncl
0.27s tests/integration/test_task.py::test_runner_uses_priority[_run_sequential]
0.26s tests/integration/test_task.py::test_base_task_names
0.25s tests/integration/test_task.py::test_run_tasks[fork-1]
0.19s tests/integration/test_task.py::test_py_diagnostic_task_constructor
0.19s tests/integration/test_task.py::test_individual_base_task_attrs
0.19s tests/integration/test_task.py::test_diagnostic_diag_script_none
```
`main` branch:
```
(esmvaltool) ✔ ~/src/esmvalgroup/esmvalcore [main|⚑ 23] 
14:46 $ pytest tests/integration/test_task.py
================================================================================================================================== test session starts ===================================================================================================================================
platform linux -- Python 3.9.1, pytest-6.2.2, py-1.10.0, pluggy-0.13.1
rootdir: /home/bandela/src/esmvalgroup/esmvalcore, configfile: setup.cfg
plugins: monitor-1.6.2, env-0.6.2, forked-1.3.0, xdist-2.2.1, flake8-1.0.7, cov-2.11.1, metadata-1.11.0, mock-3.5.1, html-3.1.1, mypy-0.8.0, anyio-3.0.1
gw0 [27] / gw1 [27]
...........................                                                                                                                                                                                                                                                        [100%]
--------------------------------------------------------------------------------------------- generated html file: file:///home/bandela/src/esmvalgroup/esmvalcore/test-reports/report.html ----------------------------------------------------------------------------------------------

----------- coverage: platform linux, python 3.9.1-final-0 -----------
Coverage HTML written to dir test-reports/coverage_html
Coverage XML written to file test-reports/coverage.xml

========================================================================================================================================== mypy ==========================================================================================================================================
Success: no issues found in 1 source file
================================================================================================================================== 27 passed in 41.18s ===================================================================================================================================
(esmvaltool) ✔ ~/src/esmvalgroup/esmvalcore [main|⚑ 23] 
14:47 $ python tests/parse_pymon.py
Looking for tests that exceed 1GB resident memory.
List of 100 slowest tests (duration, path, name
21.64s tests/integration/test_task.py::test_run_tasks[16-spawn]
11.02s tests/integration/test_task.py::test_run_tasks[4-spawn]
8.92s tests/integration/test_task.py::test_run_tasks[None-spawn]
5.51s tests/integration/test_task.py::test_run_tasks[3-spawn]
4.89s tests/integration/test_task.py::test_run_tasks[2-spawn]
3.82s tests/integration/test_task.py::test_run_tasks[spawn-2]
2.25s tests/integration/test_task.py::test_run_tasks[3-fork]
1.91s tests/integration/test_task.py::test_run_tasks[fork-2]
1.44s tests/integration/test_task.py::test_runner_uses_priority[runner1]
1.44s tests/integration/test_task.py::test_runner_uses_priority[runner1]
1.34s tests/integration/test_task.py::test_diagnostic_run_task[executable1-var0 <- "zg"]
1.33s tests/integration/test_task.py::test_diagnostic_run_task[executable2-import os\n\nprint('cow')]
1.29s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable1-diag_text1]
1.29s tests/integration/test_task.py::test_diagnostic_run_task[executable1-var0 <- "zg"]
1.29s tests/integration/test_task.py::test_run_tasks[16-fork]
1.27s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable0-diag_text0]
1.25s tests/integration/test_task.py::test_diagnostic_run_task[executable2-import os\n\nprint('cow')]
1.24s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable0-diag_text0]
1.23s tests/integration/test_task.py::test_diagnostic_run_task_fail[executable1-diag_text1]
1.22s tests/integration/test_task.py::test_diagnostic_run_task[executable0-tas = True\ntas@cow = 22\n]
1.20s tests/integration/test_task.py::test_diagnostic_run_task[executable0-tas = True\ntas@cow = 22\n]
1.16s tests/integration/test_task.py::test_run_tasks[2-fork]
1.06s tests/integration/test_task.py::test_run_tasks[4-fork]
1.04s tests/integration/test_task.py::test_run_tasks[fork-15]
0.93s tests/integration/test_task.py::test_run_tasks[fork-None]
0.91s tests/integration/test_task.py::test_run_tasks[None-fork]
0.63s tests/integration/test_task.py::test_run_tasks[1-spawn]
0.38s tests/integration/test_task.py::test_py2ncl
0.35s tests/integration/test_task.py::test_runner_uses_priority[_run_sequential]
0.27s tests/integration/test_task.py::test_runner_uses_priority[_run_sequential]
0.27s tests/integration/test_task.py::test_base_task_names
0.26s tests/integration/test_task.py::test_base_task_names
0.25s tests/integration/test_task.py::test_run_tasks[fork-1]
0.19s tests/integration/test_task.py::test_py_diagnostic_task_constructor
0.19s tests/integration/test_task.py::test_individual_base_task_attrs
0.19s tests/integration/test_task.py::test_diagnostic_diag_script_none
0.19s tests/integration/test_task.py::test_run_tasks[1-fork]
0.19s tests/integration/test_task.py::test_individual_base_task_attrs
0.19s tests/integration/test_task.py::test_py2ncl
0.19s tests/integration/test_task.py::test_diagnostic_diag_script_none
0.18s tests/integration/test_task.py::test_py_diagnostic_task_constructor
```